% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{as.BGrass_chains}
\alias{as.BGrass_chains}
\title{Combine multiple BGrass chains}
\usage{
as.BGrass_chains(BGrass_chains)
}
\arguments{
\item{BGrass_chains}{A list of objects of s3 class BGrass_chain}
}
\value{
An object of s3 class BGrass_chains
}
\description{
Combine multiple BGrass chains
}
