% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{as.BV_chains}
\alias{as.BV_chains}
\title{Combine multiple BV chains}
\usage{
as.BV_chains(BV_chains)
}
\arguments{
\item{BV_chains}{A list of objects of s3 class BV_chain}
}
\value{
An object of s3 class BV_chains
}
\description{
Combine multiple BV chains
}
