% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.BV_chain}
\alias{coef.BV_chain}
\title{Extract the logLik of the BV_chain at any position(s)}
\usage{
\method{coef}{BV_chain}(object, chain_range = NULL, ...)
}
\arguments{
\item{object}{An object of s3 class BV_chain}

\item{chain_range}{Which part of the chain should be used, by default it will use the later 80\%}

\item{...}{more arguments}
}
\value{
A list containing the estimated values of all the parameters in the chain.
}
\description{
Extract the logLik of the BV_chain at any position(s)
}
