% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getG.R
\name{getG}
\alias{getG}
\title{Get the group matrix}
\usage{
getG(AEs, data, min_size = 15)
}
\arguments{
\item{AEs}{The names of AEs}

\item{data}{A data frame that contains two columns
\itemize{
 \item{"AE"}{The names of the AE}
 \item{"group"}{The group that the AE belongs to}
}}

\item{min_size}{The minimal size of the group to be included in the matrix}
}
\value{
The group matrix G
}
\description{
Get the group matrix
}
