% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLaplacian.R
\name{getLaplacian}
\alias{getLaplacian}
\title{Get the Laplacian matrix}
\usage{
getLaplacian(AEs, data)
}
\arguments{
\item{AEs}{The names of all the adverse events}

\item{data}{A data.frame with two columns
\describe{
  \item{AE}{The name of the AE}
  \item{group}{The name of the AE group}
}}
}
\value{
The Laplacian matrix induced by the AE groups
}
\description{
Get the Laplacian matrix
}
