% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.BV_chain}
\alias{logLik.BV_chain}
\title{Extract the logLik of the BV_chain at any position(s)}
\usage{
\method{logLik}{BV_chain}(object, chain_range = object$start_pos:object$end_pos, ...)
}
\arguments{
\item{object}{An object of s3 class BV_chain}

\item{chain_range}{The range of the chain to be evaluated}

\item{...}{more arguments}
}
\value{
a vector of length \code{length(chain_range)} representing the
corresponding log likelihood.
}
\description{
Extract the logLik of the BV_chain at any position(s)
}
