% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_BGrass_chain.R
\name{sub_BGrass_chain}
\alias{sub_BGrass_chain}
\title{Extract a part of the BGrass chain}
\usage{
sub_BGrass_chain(BGrass_chain, new_start_pos, new_end_pos)
}
\arguments{
\item{BGrass_chain}{An object of s3 class BGrass_chain}

\item{new_start_pos}{New start position}

\item{new_end_pos}{New end position}
}
\value{
The truncated BGrass_chain
}
\description{
Extract a part of the BGrass chain
}
