% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_BV_chain.R
\name{sub_BV_chain}
\alias{sub_BV_chain}
\title{Extract a part of the BV chain}
\usage{
sub_BV_chain(BV_chain, new_start_pos, new_end_pos)
}
\arguments{
\item{BV_chain}{An object of s3 class BV_chain}

\item{new_start_pos}{New start position}

\item{new_end_pos}{New end position}
}
\value{
The truncated BV_chain
}
\description{
Extract a part of the BV chain
}
