% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.BV_chains}
\alias{summary.BV_chains}
\title{Extract the logLik of the BV_chain at any position(s)}
\usage{
\method{summary}{BV_chains}(object, chain_range = NULL, ...)
}
\arguments{
\item{object}{An object of s3 class BV_chain}

\item{chain_range}{Which part of the chain should be used}

\item{...}{more arguments}

\item{raw_beta}{Should the raw beta (or the marginal beta) be calculated}
}
\value{
A list containing the estimated values of all the parameters in the chain.
}
\description{
Extract the logLik of the BV_chain at any position(s)
}
