% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.BGrass_chain}
\alias{update.BGrass_chain}
\title{Update the BGrass chain}
\usage{
\method{update}{BGrass_chain}(object, n_ite = 1, max_length = Inf, ...)
}
\arguments{
\item{object}{An object of s3 class BGrass_chain}

\item{n_ite}{The number of new MCMC draws to add to the currrent chain}

\item{max_length}{The maximum length of the chain to be stored}

\item{...}{more arguments}
}
\value{
The updated BGrass_chain
}
\description{
Update the BGrass chain
}
