% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.BV_chain}
\alias{update.BV_chain}
\title{Update the BV chain}
\usage{
\method{update}{BV_chain}(object, n_ite = 1, max_length = Inf, ...)
}
\arguments{
\item{object}{An object of s3 class BV_chain}

\item{n_ite}{The number of new MCMC draws to add to the currrent chain}

\item{max_length}{The maximum length of the chain to be stored}

\item{...}{more arguments}
}
\value{
The updated BV_chain
}
\description{
Update the BV chain
}
